Updating Form.jsx
Creating the jsx for the form, also using the 'useState' hook for creating a postData object for the state to hold the values of the form fields. We reset the form fields with the reset function, and jhandle submission with the 'handleSubmit' function. Notice how we graciously changed the state of a field using the onchange prop (notice syntax)