Keeping state of the id of the post to be edited
Having the need to keep the state of the id of the post to be edited, we use the 'useState' hook to keep track of the 'currentId', and the 'setCurrentId' as the setter for the currentId. i.e setting it to null if there's no selected post to edit. Then pass these (currentId, setCurrentId) as props to the Form and Posts component. This is props drilling, & not a good approach, use REDUX instead for an elegant approach. Redux was meant for scenarios like this